+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Unreleased
+
+Added
+  * AZ-212: Additional outputs
+
 # v2.0.0 - 2019-10-15
 
 Added